dom/svg/test/test_dynamic-move-under-mouse.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=119061
-->
<head>
  <title>Test for Bug 119061</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style>
    #target:hover { fill: lime; }
  </style>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=119061">Mozilla Bug 119061</a>
<p id="display">
  <svg>
    <rect width="100%" height="100%" fill="lime"/>
    <rect id="target" width="100" height="100" transform="translate(-51,-51)" fill="lime"/>
  </svg>
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script>

SimpleTest.waitForExplicitFinish();

SimpleTest.waitForFocus(() => {
  let target = document.getElementById("target");
  let rect = document.getElementById("display").getBoundingClientRect();
  synthesizeMouse(target, 25, 25, {type: "mousemove"});
  let reference = snapshotRect(window, rect);
  target.setAttribute("fill", "red");
  ok(compareSnapshots(snapshotRect(window, rect), reference, false), "target should be visible");
  synthesizeMouse(target, 51, 51, {type: "mousemove"});
  target.removeAttribute("transform");
  ok(compareSnapshots(snapshotRect(window, rect), reference, true), "target should be invisible");

  SimpleTest.finish();
});
</script>
</pre>
</body>
</html>
